PP-6297: Move cookie banner code into cookies.js
